fix bug

+++ database/migrations/2025_10_17_103141_add_no_transaksi_to_payments_table.php
@@ -6,19 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
+    public function up(): void
+    {
+        Schema::table('payments', function (Blueprint $table) {
+            if (!Schema::hasColumn('payments', 'no_transaksi')) {
+                $table->string('no_transaksi')->nullable()->after('payment_id');
+            }
+        });
+    }
 
-public function up(): void
-{
-    Schema::table('payments', function (Blueprint $table) {
-        $table->string('no_transaksi')->nullable()->after('payment_id');
-    });
-}
-
-public function down(): void
-{
-    Schema::table('payments', function (Blueprint $table) {
-        $table->dropColumn('no_transaksi');
-    });
-}
-
+    public function down(): void
+    {
+        Schema::table('payments', function (Blueprint $table) {
+            if (Schema::hasColumn('payments', 'no_transaksi')) {
+                $table->dropColumn('no_transaksi');
+            }
+        });
+    }
 };
